api_server/requirements.txt
Django==4.0.1
django-cors-headers==3.11.0
django-storages[google]==1.12.3
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.2
google-cloud-logging==3.0.0
google-cloud-secret-manager==2.8.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
pylint==2.12.2
PyYAML==6.0
uritemplate==4.1.1
